Append minus sign to damage numbers
Healing numbers will also have a plus sign in front

+++ src/actors.h
@@ -195,7 +195,7 @@ struct EffectData {
 struct EffectState {
 	u16 lifetime;
 	u16 lifetimeCounter;
-	u8 value;
+	s16 value;
 };
 
 #ifdef EDITOR
